rename Pong to pong
keep consistency with other folder
