+++ frontend/src/components/pong/BoardGame.tsx
